[dyn-cache] reorganize
